clean lanes
