+++ metadrive/component/block/base_block.py
@@ -52,41 +52,6 @@ class BaseBlock(BaseObject, DrivableAreaProperty):
         self._block_objects = None
 
         if self.render and not self.use_render_pipeline:
-            self.ts_color = TextureStage("color")
-            self.ts_normal = TextureStage("normal")
-            self.ts_normal.setMode(TextureStage.M_normal)
-
-            # Only maintain one copy of asset
-            if self.naive_draw_map:
-                self.road_texture = self.loader.loadTexture(AssetLoader.file_path("textures", "sci", "new_color.png"))
-                self.road_normal = self.loader.loadTexture(AssetLoader.file_path("textures", "sci", "normal.jpg"))
-                self.road_texture.set_format(Texture.F_srgb)
-                self.road_normal.set_format(Texture.F_srgb)
-                self.road_texture.setMinfilter(SamplerState.FT_linear_mipmap_linear)
-                self.road_texture.setAnisotropicDegree(8)
-
-                # # continuous line
-                # self.lane_line_model = self.loader.loadModel(AssetLoader.file_path("models", "box.bam"))
-                # self.lane_line_model.setPos(0, 0, -DrivableAreaProperty.LANE_LINE_GHOST_HEIGHT / 2)
-                self.lane_line_texture = self.loader.loadTexture(AssetLoader.file_path("textures", "sci", "floor.jpg"))
-            # self.lane_line_model.setScale(DrivableAreaProperty.STRIPE_LENGTH*4,
-            #                                    DrivableAreaProperty.LANE_LINE_WIDTH,
-            #                                    DrivableAreaProperty.LANE_LINE_THICKNESS)
-            # # self.lane_line_normal = self.loader.loadTexture(
-            # #     AssetLoader.file_path("textures", "sci", "floor_normal.jpg"))
-            # # self.lane_line_texture.set_format(Texture.F_srgb)
-            # # self.lane_line_normal.set_format(Texture.F_srgb)
-            # self.lane_line_model.setTexture(self.ts_color, self.lane_line_texture)
-            # # self.lane_line_model.setTexture(self.ts_normal, self.lane_line_normal)
-            #
-            # # # broken line
-            # self.broken_lane_line_model = self.loader.loadModel(AssetLoader.file_path("models", "box.bam"))
-            # self.broken_lane_line_model.setScale(DrivableAreaProperty.STRIPE_LENGTH,
-            #                                           DrivableAreaProperty.LANE_LINE_WIDTH,
-            #                                           DrivableAreaProperty.LANE_LINE_THICKNESS)
-            # self.broken_lane_line_model.setPos(0, 0, -DrivableAreaProperty.LANE_LINE_GHOST_HEIGHT / 2)
-            # self.broken_lane_line_model.setTexture(self.ts_color, self.lane_line_texture)
-
             # side
             self.side_texture = self.loader.loadTexture(AssetLoader.file_path("textures", "sidewalk", "color.png"))
             # self.side_texture.set_format(Texture.F_srgb)
@@ -102,8 +67,6 @@ class BaseBlock(BaseObject, DrivableAreaProperty):
             self.sidewalk = self.loader.loadModel(AssetLoader.file_path("models", "box.bam"))
             self.sidewalk.setTwoSided(False)
             self.sidewalk.setTexture(self.ts_color, self.side_texture)
-            # self.sidewalk = self.loader.loadModel(AssetLoader.file_path("models", "output.egg"))
-            # self.sidewalk.setTexture(self.ts_normal, self.side_normal)
 
     def _sample_topology(self) -> bool:
         """
@@ -238,9 +201,6 @@ class BaseBlock(BaseObject, DrivableAreaProperty):
         self.lane_line_node_path.reparentTo(self.origin)
         self.lane_node_path.reparentTo(self.origin)
 
-        if self.naive_draw_map:
-            self.lane_vis_node_path.reparentTo(self.origin)
-
         try:
             self._bounding_box = self.block_network.get_bounding_box()
         except:
@@ -354,12 +314,3 @@ class BaseBlock(BaseObject, DrivableAreaProperty):
     @property
     def bounding_box(self):
         return self._bounding_box
-
-    @property
-    def naive_draw_map(self):
-        """
-        Use the most naive way to draw map
-        Returns: Boolean
-
-        """
-        return self.engine is None and self.render
