+++ metadrive/component/lane/straight_lane.py
@@ -74,17 +74,6 @@ class StraightLane(PGLane):
         self.end = end
         self.update_properties()
 
-    def construct_lane_in_block(self, block, lane_index):
-        """
-        Straight lane can be represented by one segment
-        """
-        middle = self.position(self.length / 2, 0)
-        end = self.position(self.length, 0)
-        direction_v = end - middle
-        theta = math.atan2(direction_v[1], direction_v[0])
-        width = self.width_at(0) + block.SIDEWALK_LINE_DIST * 2
-        self.construct_lane_segment(block, middle, width, self.length, theta, lane_index)
-
     @property
     def polygon(self):
         if self._polygon is None:
